minor changes + PUTMAPPING on Posts in forum

+++ src/main/java/org/example/backend_pcbuild/Community/Repository/PostCommentRepository.java
@@ -1,14 +1,20 @@
 package org.example.backend_pcbuild.Community.Repository;
 
 import org.example.backend_pcbuild.Community.Models.PostComment;
+import org.example.backend_pcbuild.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import java.util.List;
+import java.util.Optional;
 
-public interface PostCommentRepository extends JpaRepository<PostComment,Integer> {
+public interface PostCommentRepository extends JpaRepository<PostComment,Long> {
 
     @Query("SELECT pc FROM PostComment pc JOIN FETCH pc.user u WHERE pc.post.id = :postId ORDER BY pc.createdAt ASC")
     List<PostComment> findCommentsByPostId(@Param("postId") Long postId);
+
+//    Optional<PostComment> findById(Long postcommandId);
+
+//    Optional<PostComment> findByUsername(String username);
 }
